refactor(ui): split LogFilters into basic/advanced/actions
